+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - [#105](https://github.com/HewlettPackard/oneview-puppet/issues/105) Create or update uplink sets through logical interconnect groups
 - [#119](https://github.com/HewlettPackard/oneview-puppet/issues/119) Update unit tests to match updated remove_extra_unmanaged_volume from oneview-ruby-sdk
 - [#116](https://github.com/HewlettPackard/oneview-puppet/issues/116) Simplify login to i3s
+- [#121](https://github.com/HewlettPackard/oneview-puppet/issues/121) Deployment Plan and Golden Image should use the default uri parser
+- [#122](https://github.com/HewlettPackard/oneview-puppet/issues/122) Uri_parsing should support upper case for uri
 
 # 2.1.0 (2017-02-03)
 ### Version highlights: